# frontend/app/openapi-jobs.json
{
  "components": {
    "schemas": {
      "Body_photo_response_exams__exam_id__photo_response_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "student_id": {
            "title": "Student Id",
            "type": "integer"
          }
        },
        "required": [
          "student_id",
          "file"
        ],
        "title": "Body_photo_response_exams__exam_id__photo_response_post",
        "type": "object"
      },
      "Body_photo_template_exams_photo_template_post": {
        "properties": {
          "class_id": {
            "title": "Class Id",
            "type": "integer"
          },
          "exam_date": {
            "format": "date",
            "title": "Exam Date",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "default": "单元",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "file",
          "class_id",
          "name",
          "exam_date"
        ],
        "title": "Body_photo_template_exams_photo_template_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatus": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "attempts",
          "max_attempts",
          "result",
          "error",
          "created_at",
          "finished_at"
        ],
        "title": "JobStatus",
        "type": "object"
      },
      "PhotoResponseResult": {
        "properties": {
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "parse_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Job Id"
          },
          "response_id": {
            "title": "Response Id",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "response_id",
          "warnings"
        ],
        "title": "PhotoResponseResult",
        "type": "object"
      },
      "PhotoTemplateResult": {
        "properties": {
          "exam_id": {
            "title": "Exam Id",
            "type": "integer"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "parse_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Job Id"
          },
          "questions": {
            "title": "Questions",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "exam_id",
          "questions",
          "warnings"
        ],
        "title": "PhotoTemplateResult",
        "type": "object"
      },
      "QueuedJob": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "next": {
            "title": "Next",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "next"
        ],
        "title": "QueuedJob",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "DESIGN.md v0.3 MVP：知识库 -> 采集 -> 追踪 -> 归因 -> 报告",
    "title": "知行教研 · 教学质量分析平台",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/exams/photo-template": {
      "post": {
        "description": "阶段A：试卷照片 → 结构化模板 + 闭集知识点标注（source=LLM，待审核）。",
        "operationId": "photo_template_exams_photo_template_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_photo_template_exams_photo_template_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PhotoTemplateResult"
                    },
                    {
                      "$ref": "#/components/schemas/QueuedJob"
                    }
                  ],
                  "title": "Response Photo Template Exams Photo Template Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Photo Template"
      }
    },
    "/exams/{exam_id}/photo-response": {
      "post": {
        "description": "阶段B：学生卷照片 → 每题得分/选项（source=photo，状态=待审核）。",
        "operationId": "photo_response_exams__exam_id__photo_response_post",
        "parameters": [
          {
            "in": "path",
            "name": "exam_id",
            "required": true,
            "schema": {
              "title": "Exam Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_photo_response_exams__exam_id__photo_response_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PhotoResponseResult"
                    },
                    {
                      "$ref": "#/components/schemas/QueuedJob"
                    }
                  ],
                  "title": "Response Photo Response Exams  Exam Id  Photo Response Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Photo Response"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "job_status_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}